Handle edge case where user logs out mid-activity update.
If the user posts an activity update and immediately logs out, the
action to be run could switch to
`wp_async_nopriv_bp_activity_after_save` after the logout is complete.
Adding the nopriv version of the hook allows the process to continue.
Also, the action must be trimmed down before re-launching the process,
or else the hook that `handle_postback()` is attached to will be
malformed on the next load.